+++ GranjaSystem/GranjaSystem/Controllers/GeneticasController.cs
@@ -17,28 +17,43 @@ namespace GranjaSystem.Controllers
         // GET: Geneticas
         public ActionResult Index()
         {
-            return View(db.Geneticas.ToList());
+            if (Session["IdUsuario"] != null)
+            {
+                return View(db.Geneticas.ToList());
+
+            }
+            else return RedirectToAction("Index", "Login");
         }
 
         // GET: Geneticas/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Session["IdUsuario"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                tblGeneticas tblGeneticas = db.Geneticas.Find(id);
+                if (tblGeneticas == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(tblGeneticas);
+
             }
-            tblGeneticas tblGeneticas = db.Geneticas.Find(id);
-            if (tblGeneticas == null)
-            {
-                return HttpNotFound();
-            }
-            return View(tblGeneticas);
+            else return RedirectToAction("Index", "Login");
         }
 
         // GET: Geneticas/Create
         public ActionResult Create()
         {
-            return View();
+            if (Session["IdUsuario"] != null)
+            {
+
+                 return View();
+            }
+            else return RedirectToAction("Index", "Login");
         }
 
         // POST: Geneticas/Create
@@ -48,29 +63,39 @@ namespace GranjaSystem.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "IdGenetica,Genetica,Observacion")] tblGeneticas tblGeneticas)
         {
-            if (ModelState.IsValid)
+            if (Session["IdUsuario"] != null)
             {
-                db.Geneticas.Add(tblGeneticas);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                if (ModelState.IsValid)
+                {
+                    db.Geneticas.Add(tblGeneticas);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
 
-            return View(tblGeneticas);
+                return View(tblGeneticas);
+
+            }
+            else return RedirectToAction("Index", "Login");
         }
 
         // GET: Geneticas/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Session["IdUsuario"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                tblGeneticas tblGeneticas = db.Geneticas.Find(id);
+                if (tblGeneticas == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(tblGeneticas);
             }
-            tblGeneticas tblGeneticas = db.Geneticas.Find(id);
-            if (tblGeneticas == null)
-            {
-                return HttpNotFound();
-            }
-            return View(tblGeneticas);
+            else return RedirectToAction("Index", "Login");
         }
 
         // POST: Geneticas/Edit/5
@@ -80,28 +105,38 @@ namespace GranjaSystem.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "IdGenetica,Genetica,Observacion")] tblGeneticas tblGeneticas)
         {
-            if (ModelState.IsValid)
+            if (Session["IdUsuario"] != null)
             {
-                db.Entry(tblGeneticas).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+
+                if (ModelState.IsValid)
+                {
+                    db.Entry(tblGeneticas).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(tblGeneticas);
             }
-            return View(tblGeneticas);
+            else return RedirectToAction("Index", "Login");
         }
 
         // GET: Geneticas/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Session["IdUsuario"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                tblGeneticas tblGeneticas = db.Geneticas.Find(id);
+                if (tblGeneticas == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(tblGeneticas);
             }
-            tblGeneticas tblGeneticas = db.Geneticas.Find(id);
-            if (tblGeneticas == null)
-            {
-                return HttpNotFound();
-            }
-            return View(tblGeneticas);
+            else return RedirectToAction("Index", "Login");
         }
 
         // POST: Geneticas/Delete/5
@@ -109,10 +144,15 @@ namespace GranjaSystem.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            tblGeneticas tblGeneticas = db.Geneticas.Find(id);
-            db.Geneticas.Remove(tblGeneticas);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Session["IdUsuario"] != null)
+            {
+                tblGeneticas tblGeneticas = db.Geneticas.Find(id);
+                db.Geneticas.Remove(tblGeneticas);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+
+            }
+            else return RedirectToAction("Index", "Login");
         }
 
         protected override void Dispose(bool disposing)
